vit v0.1 unable to load checkpoint, runnable

+++ mask2former/modeling/criterion.py
@@ -161,7 +161,7 @@ class SetCriterion(nn.Module):
         with torch.no_grad():
             # sample point_coords
             point_coords = get_uncertain_point_coords_with_randomness(
-                src_masks,
+                src_masks.float(),
                 lambda logits: calculate_uncertainty(logits),
                 self.num_points,
                 self.oversample_ratio,
@@ -169,13 +169,13 @@ class SetCriterion(nn.Module):
             )
             # get gt labels
             point_labels = point_sample(
-                target_masks,
+                target_masks.float(),
                 point_coords,
                 align_corners=False,
             ).squeeze(1)
 
         point_logits = point_sample(
-            src_masks,
+            src_masks.float(),
             point_coords,
             align_corners=False,
         ).squeeze(1)
